--- Assets/Oculus/VR/Editor/Tools.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a210e7e0f4c269042aa589944b3c1c9e
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
