--- Assets/Plugins.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ff3f10f840c68844793d43fcebcbc418
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
