--- Assets/Prefab/SongListButton.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3f62fe1308447b7449a016c80c5e455c
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
