--- Assets/Scripts/Menu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 570f54418b6848f4eb7595af411f8af4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
