--- Assets/Scripts/Menu/SongListButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4e5e5603936364a45a0fbf32861822d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
