--- Assets/Scripts/Menu/SongListController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ffb38c56c03db4a48a2db8c40ab23f47
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
